lxk更改

+++ face/InsightFace_Pytorch-master/demo.py
@@ -31,7 +31,7 @@ device = 'cuda'
 # print(img1.shape)
 
 # 指定保存位置
-save_path = "C:/Users/10698/Desktop/FaceCompare-master/FaceCompare-master/face"
+save_path = "C:/Users/10698/Desktop/ff/FaceCompare/face"
 file_path = os.path.join(save_path, "uploaded_image.jpg")
 
 #被测试图片
@@ -47,12 +47,16 @@ emb1 = model(img1)[0]
 print(emb1.shape)
 
 #读取图片库
-folder_path = 'C:/Users/10698/Desktop/FaceCompare-master/FaceCompare-master/webui/static/test/'
+folder_path = 'C:/Users/10698/Desktop/ff/FaceCompare/webui/static/test/'
 
-#满足条件的图片库
-destination_folder = 'C:/Users/10698/Desktop/FaceCompare-master/FaceCompare-master/webui/static/manzu/'
-# 删除文件夹及其内容
-shutil.rmtree(destination_folder)
+
+# 满足条件的图片库
+destination_folder = 'C:/Users/10698/Desktop/ff/FaceCompare/webui/static/manzu/'
+
+# 检查文件夹是否存在
+if os.path.exists(destination_folder):
+    # 删除文件夹及其内容
+    shutil.rmtree(destination_folder)
 
 # 重新创建文件夹
 os.makedirs(destination_folder)
